ingestion-api/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-multipart
google-genai
